backend/requirements.txt
# Web框架
Flask==2.3.3
Werkzeug==2.3.7
Flask-CORS==4.0.0

# WebSocket
gevent==23.9.1
gevent-websocket==0.10.1
websockets==12.0

# 数据库
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
redis==5.0.1
alembic==1.13.0

# MT5集成
MetaTrader5==5.0.44

# 数据分析
pandas==2.1.4
numpy==1.25.2
pandas-ta==0.3.14
scipy==1.11.3
PyPortfolioOpt==1.5.5
statsmodels==0.14.0

# 工具库
python-dotenv==1.0.0
pydantic==2.4.0
marshmallow==3.20.1
requests==2.31.0

# 异步支持
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# 日志和监控
loguru==0.7.2
prometheus-client==0.19.0